requirements.txt
fastapi
uvicorn
requests
pandas
k-means-constrained
python-dotenv
